+++ cypress/e2e/inventory/holdings/holdings-record-through-upload.cy.js
@@ -6,7 +6,7 @@ import HoldingsRecordView from '../../../support/fragments/inventory/holdingsRec
 import testTypes from '../../../support/dictionary/testTypes';
 import features from '../../../support/dictionary/features';
 import DataImport from '../../../support/fragments/data_import/dataImport';
-import SearchInventory from '../../../support/fragments/data_import/searchInventory';
+import InventorySearch from '../../../support/fragments/inventory/inventorySearch';
 import InventoryInstances from '../../../support/fragments/inventory/inventoryInstances';
 import NewHoldingsRecord from '../../../support/fragments/inventory/newHoldingsRecord';
 import HoldingsRecordEdit from '../../../support/fragments/inventory/holdingsRecordEdit';
@@ -22,7 +22,7 @@ describe('Manage holding records of instance records created through marc file u
   it('C345408 MARC instance record + FOLIO holdings record (Regression) (spitfire)', { tags: [testTypes.smoke, DevTeams.spitfire, features.holdingsRecord, testTypes.broken] }, () => {
     DataImport.uploadMarcBib().then(instanceRecordHrId => {
       cy.visit(TopMenu.inventoryPath);
-      SearchInventory.searchInstanceByHRID(instanceRecordHrId);
+      InventorySearch.searchInstanceByHRID(instanceRecordHrId);
       InventoryInstances.selectInstance();
       InventoryInstance.waitLoading();
       InventoryInstance.createHoldingsRecord();
