--- cypress/support/fragments/data_import/searchInventory.js
@@ -1,91 +1,0 @@
-import {
-  Button,
-  KeyValue,
-  MultiColumnListCell,
-  TextField,
-  Select
-} from '../../../../interactors';
-import logsViewAll from './logs/logsViewAll';
-import DateTools from '../../utils/dateTools';
-import InventoryInstances from '../inventory/inventoryInstances';
-
-// TODO replace to Inventory/inventorySearch
-const getInstanceHRID = () => {
-  return logsViewAll.getSingleJobProfile() // get the first job id from job logs list
-    .then(({ id }) => {
-      // then, make request with the job id
-      // and get the only record id inside the uploaded file
-      const queryString = 'limit=1000&order=asc';
-      return cy.request({
-        method: 'GET',
-        url: `${Cypress.env('OKAPI_HOST')}/metadata-provider/jobLogEntries/${id}?${queryString}`,
-        headers: {
-          'x-okapi-tenant': Cypress.env('OKAPI_TENANT'),
-          'x-okapi-token': Cypress.env('token'),
-        },
-      })
-        .then(({ body: { entries } }) => {
-          // then, make request with the job id and the record id
-          // and get Instance HRID
-          const recordId = entries[0].sourceRecordId;
-          return cy.request({
-            method: 'GET',
-            url: `${Cypress.env('OKAPI_HOST')}/metadata-provider/jobLogEntries/${id}/records/${recordId}`,
-            headers: {
-              'x-okapi-tenant': Cypress.env('OKAPI_TENANT'),
-              'x-okapi-token': Cypress.env('token'),
-            },
-          })
-            .then(({ body: { relatedInstanceInfo } }) => {
-              return relatedInstanceInfo.hridList;
-            });
-        });
-    });
-};
-
-const searchInstanceByHRID = (id) => {
-  InventoryInstances.waitContentLoading();
-  cy.do([
-    Select({ id: 'input-inventory-search-qindex' }).choose('Instance HRID'),
-    TextField({ id: 'input-inventory-search' }).fillIn(id),
-    Button('Search').click()
-  ]);
-  InventoryInstances.waitLoading();
-};
-
-const searchInstanceByTitle = (title) => {
-  cy.do([
-    TextField({ id: 'input-inventory-search' }).fillIn(title),
-    Button('Search').click()
-  ]);
-  InventoryInstances.waitLoading();
-};
-
-// when creating mapping profile we choose status cataloged date as today
-// in inventory, it will be in YYYY-MM-DD format
-const expectedCatalogedDate = DateTools.getFormattedDate({ date: new Date() });
-// when creating mapping profile we choose instance status term as "Batch Loaded"
-// in inventory, this will be "batch" for status code and "Batch Loaded" for status term
-const expectedStatusTerm = 'Batch Loaded';
-const expectedStatusCode = 'batch';
-
-const checkInstanceDetails = () => {
-  cy.do(MultiColumnListCell({ row: 0, columnIndex: 1 }).click());
-  const catalogedDate = KeyValue('Cataloged date');
-  const instanceStatusTerm = KeyValue('Instance status term');
-  const instanceStatusCode = KeyValue('Instance status code');
-
-  cy.expect(catalogedDate.has({ value: expectedCatalogedDate }));
-  cy.expect(instanceStatusTerm.has({ value: expectedStatusTerm }));
-  cy.expect(instanceStatusCode.has({ value: expectedStatusCode }));
-};
-
-export default {
-  // TODO: move to existing InventorySearch fragment
-  getInstanceHRID,
-  // TODO: move to existing InventorySearch fragment
-  searchInstanceByHRID,
-  // TODO: move to existing  inventory fragment
-  checkInstanceDetails,
-  searchInstanceByTitle,
-};
