+++ cypress/support/fragments/inventory/inventorySearch.js
@@ -12,8 +12,12 @@ import {
   Select,
   Form,
   TextInput,
+  KeyValue
 } from '../../../../interactors';
 import InventoryActions from './inventoryActions';
+import InventoryInstances from './inventoryInstances';
+import logsViewAll from '../data_import/logs/logsViewAll';
+import DateTools from '../../utils/dateTools';
 
 const effectiveLocationInput = Accordion({ id: 'effectiveLocation' });
 const languageInput = Accordion({ id: 'language' });
@@ -22,9 +26,86 @@ const searchButton = Button('Search');
 const searchTextField = TextField('Search ');
 const inventorySearch = TextInput({ id: 'input-inventory-search' });
 
+const searchInstanceByHRID = (id) => {
+  InventoryInstances.waitContentLoading();
+  cy.do([
+    Select({ id: 'input-inventory-search-qindex' }).choose('Instance HRID'),
+    TextField({ id: 'input-inventory-search' }).fillIn(id),
+    Button('Search').click()
+  ]);
+  InventoryInstances.waitLoading();
+};
+
+const searchInstanceByTitle = (title) => {
+  cy.do([
+    TextField({ id: 'input-inventory-search' }).fillIn(title),
+    Button('Search').click()
+  ]);
+  InventoryInstances.waitLoading();
+};
+
+const getInstanceHRID = () => {
+  return logsViewAll.getSingleJobProfile() // get the first job id from job logs list
+    .then(({ id }) => {
+      // then, make request with the job id
+      // and get the only record id inside the uploaded file
+      const queryString = 'limit=1000&order=asc';
+      return cy.request({
+        method: 'GET',
+        url: `${Cypress.env('OKAPI_HOST')}/metadata-provider/jobLogEntries/${id}?${queryString}`,
+        headers: {
+          'x-okapi-tenant': Cypress.env('OKAPI_TENANT'),
+          'x-okapi-token': Cypress.env('token'),
+        },
+      })
+        .then(({ body: { entries } }) => {
+          // then, make request with the job id and the record id
+          // and get Instance HRID
+          const recordId = entries[0].sourceRecordId;
+          return cy.request({
+            method: 'GET',
+            url: `${Cypress.env('OKAPI_HOST')}/metadata-provider/jobLogEntries/${id}/records/${recordId}`,
+            headers: {
+              'x-okapi-tenant': Cypress.env('OKAPI_TENANT'),
+              'x-okapi-token': Cypress.env('token'),
+            },
+          })
+            .then(({ body: { relatedInstanceInfo } }) => {
+              return relatedInstanceInfo.hridList;
+            });
+        });
+    });
+};
+
+const checkInstanceDetails = () => {
+  // when creating mapping profile we choose status cataloged date as today
+  // in inventory, it will be in YYYY-MM-DD format
+  const expectedCatalogedDate = DateTools.getFormattedDate({ date: new Date() });
+  // when creating mapping profile we choose instance status term as "Batch Loaded"
+  // in inventory, this will be "batch" for status code and "Batch Loaded" for status term
+  const expectedStatusTerm = 'Batch Loaded';
+  const expectedStatusCode = 'batch';
+
+  cy.do(MultiColumnListCell({ row: 0, columnIndex: 1 }).click());
+  const catalogedDate = KeyValue('Cataloged date');
+  const instanceStatusTerm = KeyValue('Instance status term');
+  const instanceStatusCode = KeyValue('Instance status code');
+
+  cy.expect(catalogedDate.has({ value: expectedCatalogedDate }));
+  cy.expect(instanceStatusTerm.has({ value: expectedStatusTerm }));
+  cy.expect(instanceStatusCode.has({ value: expectedStatusCode }));
+};
+
 export default {
+  searchInstanceByHRID,
+  searchInstanceByTitle,
+  getInstanceHRID,
+  checkInstanceDetails,
   getAllSearchResults: () => MultiColumnList(),
   getSearchResult: (row = 0, col = 0) => MultiColumnListCell({ 'row': row, 'columnIndex': col }),
+  waitLoading: () => cy.expect([Form().find(inventorySearch).exists()]),
+  browseCallNumberIsAbsent:() => cy.expect(HTML('Browse call numbers').absent()),
+  browseSubjectIsAbsent:() => cy.expect(HTML('Browse subjects').absent()),
 
   effectiveLocation: {
     mainLibrary: { id: 'clickable-filter-effectiveLocation-main-library' }
@@ -32,11 +113,6 @@ export default {
 
   language: {
     eng: { id: 'clickable-filter-language-english' }
-  },
-  waitLoading: () => {
-    cy.expect([
-      Form().find(inventorySearch).exists(),
-    ]);
   },
 
   selectResultCheckboxes(count) {
@@ -85,14 +161,6 @@ export default {
     // eslint-disable-next-line cypress/no-unnecessary-waiting
     cy.wait(1000);
     cy.do(Select('Search field index').choose('Browse subjects'));
-  },
-
-  browseCallNumberIsAbsent() {
-    cy.expect(HTML('Browse call numbers').absent());
-  },
-
-  browseSubjectIsAbsent() {
-    cy.expect(HTML('Browse subjects').absent());
   },
 
   showsOnlyEffectiveLocation() {
@@ -179,12 +247,8 @@ export default {
   searchByParameter: (parameter, value) => {
     cy.do(SearchField({ id: 'input-inventory-search' }).selectIndex(parameter));
     cy.do(searchTextField.fillIn(value));
-    // TODO: clarify the reason of failed waiter
-    // cy.intercept('/holdings-storage/holdings?*').as('getHoldings');
-    // cy.intercept('/copycat/profiles?*').as('getProfiles');
     cy.do(searchButton.focus());
     cy.do(searchButton.click());
-    // cy.wait(['@getHoldings', '@getProfiles']);
   },
   simpleSearchByParameter: (parameter, value) => {
     cy.do(SearchField({ id: 'input-inventory-search' }).selectIndex(parameter));
@@ -198,17 +262,9 @@ export default {
     cy.do(searchButton.focus());
     cy.do(searchButton.click());
   },
-  switchToItem: () => {
-    cy.do(Button({ id: 'segment-navigation-items' }).click());
-  },
-
-  switchToHoldings: () => {
-    cy.do(Button({ id: 'segment-navigation-holdings' }).click());
-  },
-
-  switchToInstance: () => {
-    cy.do(Button({ id: 'segment-navigation-instances' }).click());
-  },
+  switchToItem: () => cy.do(Button({ id: 'segment-navigation-items' }).click()),
+  switchToHoldings: () => cy.do(Button({ id: 'segment-navigation-holdings' }).click()),
+  switchToInstance: () => cy.do(Button({ id: 'segment-navigation-instances' }).click()),
 
   instanceTabIsDefault() {
     cy.do(
@@ -226,9 +282,7 @@ export default {
     cy.expect(Pane({ id:'pane-results' }).find(MultiColumnListHeader()).exists());
   },
 
-  verifySearchResult(cellContent) {
-    cy.expect(MultiColumnListCell({ content: cellContent }).exists());
-  },
+  verifySearchResult:(cellContent) => cy.expect(MultiColumnListCell({ content: cellContent }).exists()),
 
   getInstancesByIdentifierViaApi(identifier, limit = 100) {
     return cy
